make login

+++ app/src/main/java/ru/delayvi/drive3/domain/entity/users/User.kt
@@ -3,11 +3,11 @@ package ru.delayvi.drive3.domain.entity.users
 import java.util.*
 
 data class User(
-    val id: Int,
+    val id: Long,
     var username: String,
 ) {
     companion object {
-        val id: Int? = null
+        val id: Long? = null
         val username: String? = null
     }
